// coair-designer-pack/src/types/api.ts
export interface RelatedDoc {
  doc_id: string;
  doc_name: string;
  date: string;
  doc_type: string;
  reason: string;
  score: number | null;
  sender: string;
  recipient: string;
}